+++ src/Webform.unit.js
@@ -15,6 +15,7 @@ import {
   manualOverride,
   validationOnBlur,
   calculateValueWithManualOverride,
+  calculateValueWithSubmissionMetadata,
   formWithAdvancedLogic,
   formWithPatternValidation,
   calculatedSelectboxes,
@@ -34,6 +35,7 @@ import {
   formWithDateTimeComponents,
   formWithCollapsedPanel,
   formWithCustomFormatDate,
+  tooltipActivateCheckbox,
 } from '../test/formtest';
 // import DataGridOnBlurValidation from '../test/forms/dataGridOnBlurValidation';
 import UpdateErrorClassesWidgets from '../test/forms/updateErrorClasses-widgets';
@@ -48,7 +50,8 @@ import formWithDataGridWithCondColumn from '../test/forms/dataGridWithConditiona
 import { nestedFormInWizard } from '../test/fixtures';
 import NativePromise from 'native-promise-only';
 import { fastCloneDeep } from '../lib/utils/utils';
-
+import dataGridOnBlurValidation from '../test/forms/dataGridOnBlurValidation';
+import checkBlurFocusEventForm from '../test/forms/checkBlurFocusEventForm';
 import truncateMultipleSpaces from '../test/forms/truncateMultipleSpaces';
 import calculatedValue from '../test/forms/calculatedValue';
 import conditionalDataGridWithTableAndRadio from '../test/forms/conditionalDataGridWithTableAndRadio';
@@ -63,11 +66,14 @@ import formWithCalcValue from '../test/forms/formWithCalcValue';
 import formWithAllowCalculateOverride from '../test/forms/formWithAllowCalculateOverride';
 import testClearOnHideInsideEditGrid from '../test/forms/clearOnHideInsideEditGrid';
 import formWithNestedDataGridInitEmpty from '../test/forms/nestedDataGridWithInitEmpty';
+import formWithEventLogicInHiddenComponent from '../test/forms/formWithEventLogicInHiddenComponent';
 import * as FormioUtils from './utils/utils';
 import htmlRenderMode from '../test/forms/htmlRenderMode';
 import optionalSanitize from '../test/forms/optionalSanitize';
 import formWithCheckboxRadioaType from '../test/forms/formWithCheckboxRadioType';
+import formsWithNewSimpleConditions from '../test/forms/formsWithNewSimpleConditions';
 import formWithRadioInsideDataGrid from '../test/forms/formWithRadioInsideDataGrid';
+import formWithCheckboxRadioType from '../test/forms/formWithCheckboxRadioType';
 
 global.requestAnimationFrame = (cb) => cb();
 global.cancelAnimationFrame = () => {};
@@ -88,8 +94,55 @@ describe('Webform tests', function() {
     }).catch((err) => done(err));
   });
 
-  it('Should return correct strign value for checkbox radio type', function(done) {
-    Formio.createForm(formWithCheckboxRadioaType).then((form) => {
+  it('Should fire blur and focus events for address and select components', function(done) {
+    const formElement = document.createElement('div');
+    const form = new Webform(formElement);
+
+    form.setForm(checkBlurFocusEventForm).then(() => {
+      let blurEvents = 0;
+      let focusEvents = 0;
+      form.on('blur', () => {
+        blurEvents = blurEvents + 1;
+      });
+
+      form.on('focus', () => {
+        focusEvents = focusEvents + 1;
+      });
+
+      const focusEvent = new Event('focus');
+      const blurEvent = new Event('blur');
+
+      const selectChoices = form.getComponent('selectChoices');
+      selectChoices.focusableElement.dispatchEvent(focusEvent);
+
+      setTimeout(() => {
+        selectChoices.focusableElement.dispatchEvent(blurEvent);
+
+        const selectHtml = form.getComponent('selectHtml');
+        selectHtml.refs.selectContainer.dispatchEvent(focusEvent);
+
+        setTimeout(() => {
+          selectHtml.refs.selectContainer.dispatchEvent(blurEvent);
+
+          const address = form.getComponent('address');
+          address.refs.searchInput[0].dispatchEvent(focusEvent);
+
+          setTimeout(() => {
+            address.refs.searchInput[0].dispatchEvent(blurEvent);
+
+            setTimeout(() => {
+              assert.equal(focusEvents, 3);
+              assert.equal(blurEvents, 3);
+              done();
+            }, 300);
+          }, 300);
+        }, 300);
+      }, 300);
+    }).catch((err) => done(err));
+  });
+
+  it('Should return correct string value for checkbox radio type', function(done) {
+    Formio.createForm(formWithCheckboxRadioType).then((form) => {
       form.setValue({ data: { radio: 'value1', checkbox: true } });
       setTimeout(() => {
         const stringValues = {
@@ -121,6 +174,38 @@ describe('Webform tests', function() {
     }).catch((err) => done(err));
   });
 
+  it('Should set value for hidden nested component through the logic triggered by event', function(done) {
+    const formElement = document.createElement('div');
+    const form = new Webform(formElement);
+
+    form.setForm(formWithEventLogicInHiddenComponent).then(() => {
+      const regesteredAddress = form.getComponent('registeredAddressInformation').getComponent('streetAddress')[0];
+      const address =  form.getComponent('addressInformation').getComponent('streetAddress')[0];
+
+      assert.equal(address.visible, true);
+      assert.equal(regesteredAddress.visible, false);
+
+      const value = 'Dallas';
+      address.setValue(value);
+
+      setTimeout(() => {
+        assert.equal(address.dataValue, value);
+        assert.equal(regesteredAddress.dataValue, value);
+
+        const role =  form.getComponent('role');
+        role.setValue(['client']);
+
+        setTimeout(() => {
+          assert.equal(address.visible, false);
+          assert.equal(regesteredAddress.visible, true);
+          assert.equal(regesteredAddress.dataValue, value);
+          assert.equal(address.dataValue, value);
+          done();
+        }, 500);
+      }, 500);
+    }).catch((err) => done(err));
+  });
+
   it('Should recalculate value when submission is being set in edit mode', function(done) {
     const formElement = document.createElement('div');
     const form = new Webform(formElement);
@@ -143,6 +228,22 @@ describe('Webform tests', function() {
         }, 500);
       });
     }).catch((err) => done(err));
+  });
+
+  it('Should not activate checkbox when clicking tooltip icon', function(done) {
+    const element = document.createElement('div');
+    const form = new Webform(element);
+
+    form.setForm(tooltipActivateCheckbox).then(() => {
+    const checkboxValue = form.element.querySelector('[name="data[checkbox]"]').value;
+    Harness.clickElement(form, form.element.querySelector('[ref="tooltip"]'));
+
+    setTimeout(() => {
+      assert.equal(form.element.querySelector('[name="data[checkbox]"]').value, checkboxValue);
+      done();
+    }, 200);
+    })
+    .catch((err) => done(err));
   });
 
   it('Should show survey values in html render mode', function(done) {
@@ -1660,7 +1761,7 @@ describe('Webform tests', function() {
     }).catch(done);
   };
 
-  //TOFIX
+  //BUG - uncomment once fixed (ticket FIO-6042)
   // it('Should not fire validations when fields are either protected or not persistent.', (done) => {
   //   const form = new Webform(formElement,{ language: 'en', template: 'bootstrap3' });
   //   form.setForm(
@@ -2162,27 +2263,27 @@ describe('Webform tests', function() {
       }).catch(done);
     });
 
-    //TOFIX
-    // it('Should keep components inside DataGrid valid onChange', (done) => {
-    //   formElement.innerHTML = '';
-    //   const form = new Webform(formElement, { language: 'en', template: 'bootstrap3' });
-    //   form.setForm(DataGridOnBlurValidation).then(() => {
-    //     const component = form.components[0];
-    //     Harness.setInputValue(component, 'data[dataGrid][0][textField]', '12');
-    //     const textField = component.iteratableRows[0].components.textField;
-    //     setTimeout(() => {
-    //       assert.equal(textField.error, '', 'Should stay valid on input');
-    //       const blur = new Event('blur', { bubbles: true, cancelable: true });
-    //       const input = textField.refs.input[0];
-    //       input.dispatchEvent(blur);
-    //       textField.element.dispatchEvent(blur);
-    //         setTimeout(() => {
-    //           assert(textField.error, 'Should be validated after blur');
-    //           done();
-    //         }, 250);
-    //     }, 250);
-    //   }).catch(done);
-    // });
+    it('Should keep components inside DataGrid valid onChange', (done) => {
+      formElement.innerHTML = '';
+      const form = new Webform(formElement, { language: 'en', template: 'bootstrap3' });
+      form.setForm(dataGridOnBlurValidation).then(() => {
+        const component = form.components[0];
+        Harness.setInputValue(component, 'data[dataGrid][0][textField]', '12');
+
+        setTimeout(() => {
+          const textField = component.iteratableRows[0].components.textField;
+          assert.equal(!!textField.error, false, 'Should stay valid on input');
+          const blur = new Event('blur', { bubbles: true, cancelable: true });
+          const input = textField.refs.input[0];
+          input.dispatchEvent(blur);
+          textField.element.dispatchEvent(blur);
+            setTimeout(() => {
+              assert(textField.error, 'Should be validated after blur');
+              done();
+            }, 250);
+        }, 250);
+      }).catch(done);
+    });
   });
 
   describe('Reset values', () => {
@@ -2287,6 +2388,270 @@ describe('Webform tests', function() {
     });
   });
 
+  describe('New Simple Conditions', () => {
+    it('Should show field if all conditions are met', function(done) {
+      const formElement = document.createElement('div');
+      const form = new Webform(formElement);
+
+      form.setForm(formsWithNewSimpleConditions.form1).then(() => {
+        const conditionalComponent = form.getComponent('conditionalField');
+        assert.equal(conditionalComponent.visible, false, '(1) Component should be conditionally hidden');
+
+        form.setValue({ data: { number: 11, email: 'test@form.io', radio: 'one' } });
+
+        setTimeout(() => {
+          assert.equal(conditionalComponent.visible, true, '(2) Component should be conditionally visible');
+          const emailComponent = form.getComponent('email');
+
+          emailComponent.setValue('test@form1.io');
+
+          setTimeout(() => {
+            assert.equal(conditionalComponent.visible, false, '(3) Component should be conditionally hidden');
+            done();
+          }, 300);
+        }, 300);
+      }).catch((err) => done(err));
+    });
+
+    it('Should show field if any condition is met', function(done) {
+      const formElement = document.createElement('div');
+      const form = new Webform(formElement);
+      const formCopy = fastCloneDeep(formsWithNewSimpleConditions.form1);
+      _.set(formCopy, 'components[0].conditional.conjunction', 'any');
+
+      form.setForm(formCopy).then(() => {
+        const conditionalComponent = form.getComponent('conditionalField');
+        assert.equal(conditionalComponent.visible, false, '(1) Component should be conditionally hidden');
+
+        form.setValue({ data: { number: 1100, email: 'test@form.io' } });
+
+        setTimeout(() => {
+          assert.equal(conditionalComponent.visible, true, '(2) Component should be conditionally visible');
+          form.setValue({ data: { number: 10 , email: 'test@form1.io' } });
+
+          setTimeout(() => {
+            assert.equal(conditionalComponent.visible, true, '(3) Component should be conditionally visible');
+            form.setValue({ data: { number: 10000 } });
+
+            setTimeout(() => {
+              assert.equal(conditionalComponent.visible, false, '(4) Component should be conditionally hidden');
+              form.setValue({ data: { radio: 'one' } });
+
+              setTimeout(() => {
+                assert.equal(conditionalComponent.visible, true, '(5) Component should be conditionally visible');
+
+                done();
+              }, 450);
+            }, 400);
+          }, 350);
+        }, 300);
+      }).catch((err) => done(err));
+    });
+
+    it('Should hide field if any condition is met', function(done) {
+      const formElement = document.createElement('div');
+      const form = new Webform(formElement);
+      const formCopy = fastCloneDeep(formsWithNewSimpleConditions.form1);
+      _.set(formCopy, 'components[0].conditional.show', false);
+      _.set(formCopy, 'components[0].conditional.conjunction', 'any');
+
+      form.setForm(formCopy).then(() => {
+        const conditionalComponent = form.getComponent('conditionalField');
+        assert.equal(conditionalComponent.visible, true, 'Component should be conditionally visible');
+
+        form.setValue({ data: { number: 1100, email: 'test@form.io' } });
+
+        setTimeout(() => {
+          assert.equal(conditionalComponent.visible, false, 'Component should be conditionally hidden');
+          form.setValue({ data: { number: 10 , email: 'test@form1.io' } });
+
+          setTimeout(() => {
+            assert.equal(conditionalComponent.visible, false, 'Component should be conditionally hidden');
+            form.setValue({ data: { number: 10000 } });
+
+            setTimeout(() => {
+              assert.equal(conditionalComponent.visible, true, 'Component should be conditionally visible');
+              form.setValue({ data: { radio: 'one' } });
+
+              setTimeout(() => {
+                assert.equal(conditionalComponent.visible, false, 'Component should be conditionally hidden');
+                done();
+              }, 300);
+            }, 300);
+          }, 300);
+        }, 300);
+      }).catch((err) => done(err));
+    });
+
+    it('Should hide field if all conditions are met', function(done) {
+      const formElement = document.createElement('div');
+      const form = new Webform(formElement);
+      const formCopy = fastCloneDeep(formsWithNewSimpleConditions.form1);
+      _.set(formCopy, 'components[0].conditional.show', false);
+
+      form.setForm(formCopy).then(() => {
+        const conditionalComponent = form.getComponent('conditionalField');
+        assert.equal(conditionalComponent.visible, true, 'Component should be conditionally visible');
+
+        form.setValue({ data: { number: 11, email: 'test@form.io', radio: 'one' } });
+
+        setTimeout(() => {
+          assert.equal(conditionalComponent.visible, false, 'Component should be conditionally hidden');
+          const emailComponent = form.getComponent('email');
+
+          emailComponent.setValue('test@form1.io');
+
+          setTimeout(() => {
+            assert.equal(conditionalComponent.visible, true, 'Component should be conditionally visible');
+            done();
+          }, 300);
+        }, 300);
+      }).catch((err) => done(err));
+    });
+
+    it('Should show field if all conditions are met (test with different component types + multiple components)', function(done) {
+      const formElement = document.createElement('div');
+      const form = new Webform(formElement);
+
+      form.setForm(formsWithNewSimpleConditions.form2).then(() => {
+        const conditionalComponent = form.getComponent('conditionalField');
+        assert.equal(conditionalComponent.visible, false, 'Component should be conditionally hidden');
+
+        form.setValue({
+          data: {
+            email: 'test@form.io',
+            day: '10/06/2022',
+            survey: {
+              q1: 'true',
+            },
+            number: [100, 25, 350],
+            checkbox: true,
+            selectBoxes: {
+              one: true,
+              two: false,
+              three: false,
+              four: false,
+              five: false,
+            },
+            radio: 'two',
+            tags: 'test,newtag',
+            selectValues: 'one',
+            selectCustomWithValuesOfNumberType: 1,
+            submit: true,
+            currency: 35,
+          }
+        });
+
+        setTimeout(() => {
+          assert.equal(conditionalComponent.visible, true, 'Component should be conditionally visible');
+          const dayComponent = form.getComponent('day');
+
+          dayComponent.setValue('8/09/2022');
+
+          setTimeout(() => {
+            assert.equal(conditionalComponent.visible, false, 'Component should be conditionally hidden');
+            done();
+          }, 300);
+        }, 300);
+      }).catch((err) => done(err));
+    });
+
+    it('Should show/hide field inside datagrid rows', function(done) {
+      const formElement = document.createElement('div');
+      const form = new Webform(formElement);
+
+      form.setForm(formsWithNewSimpleConditions.form4).then(() => {
+        const dataGrid = form.getComponent('dataGrid');
+
+        dataGrid.setValue([
+          { number: 50 },
+          { number: 55 },
+          { number: 12 },
+          { number: 105 },
+        ]);
+
+        setTimeout(() => {
+          const expectedValues = {
+            '0': true,
+            '1': false,
+            '2': true,
+            '3': false
+          };
+
+          _.each(dataGrid.rows, (row, index) => {
+            assert.equal(row['textField'].visible, expectedValues[`${index}`], `Component should be conditionally ${expectedValues[`${index}`] ? 'visible': 'hidden'} in row ${index}`);
+          });
+          done();
+        }, 300);
+      }).catch((err) => done(err));
+    });
+
+    it('Should set component value through logic triggered by simple condition', function(done) {
+      const formElement = document.createElement('div');
+      const form = new Webform(formElement);
+
+      form.setForm(formsWithNewSimpleConditions.form3).then(() => {
+        const componentWithLogic = form.getComponent('fieldWithLogic');
+        assert.equal(componentWithLogic.isEmpty(), true, 'Component should be empty');
+
+        form.setValue({
+          data: {
+            number: 2,
+            radio: 'two'
+          }
+        });
+
+        setTimeout(() => {
+          assert.equal(componentWithLogic.dataValue, 'logic works', 'Component should have value set by logic');
+          done();
+        }, 300);
+      }).catch((err) => done(err));
+    });
+
+    it('Should show field if all conditions are met (test all operators)', function(done) {
+      const formElement = document.createElement('div');
+      const form = new Webform(formElement);
+
+      form.setForm(formsWithNewSimpleConditions.form5).then(() => {
+        const conditionalComponent = form.getComponent('conditionalField');
+        assert.equal(conditionalComponent.visible, false, '(1) Component should be conditionally hidden');
+
+        form.setValue({ data: {
+          dateTime: '2022-09-29T12:00:00+03:00',
+          day: '09/29/2022',
+          dateTime1: '2022-09-29T12:00:00+03:00',
+          day1: '09/29/2022',
+          url: 'portal.form.io',
+          number: 100,
+          currency: 100,
+          textField1: 'some test text',
+          day2: '09/29/2022',
+          select: '',
+          radio: 'one',
+          dateTime3: '2022-09-12T12:00:00+03:00',
+          textArea: 'test',
+          textField2: 'test2',
+          number2: [100],
+          currency2: 100,
+          email: 'some@form.io',
+          url2: 'portal.form.io',
+        } });
+
+        setTimeout(() => {
+          assert.equal(conditionalComponent.visible, true, '(2) Component should be conditionally visible');
+          const selectComponent = form.getComponent('select');
+
+          selectComponent.setValue('one');
+
+          setTimeout(() => {
+            assert.equal(conditionalComponent.visible, false, '(3) Component should be conditionally hidden');
+            done();
+          }, 300);
+        }, 300);
+      }).catch((err) => done(err));
+    });
+  });
+
   describe('Calculate Value with allowed manual override', () => {
     const initialSubmission = {
       data: {
@@ -2352,6 +2717,30 @@ describe('Webform tests', function() {
               });
             }, 250);
           }, 250);
+        }, 250);
+      }).catch(done);
+    });
+
+    it('Should apply submission metadata value in calculation.', (done) => {
+      const formElement = document.createElement('div');
+      const form = new Webform(formElement, { language: 'en', template: 'bootstrap3' });
+      form.setForm(calculateValueWithSubmissionMetadata).then(() => {
+        const textField = form.getComponent('textField');
+
+        textField.setValue('test value');
+
+        form.submit(false, {});
+
+        setTimeout(() => {
+          expect(form.submission.metadata).to.exist;
+          expect(form.submission.metadata.timezone).to.be.not.empty;
+          expect(form.submission.data.textField).to.be.not.empty;
+          expect(form.submission.data.textArea).to.be.not.empty;
+          expect(form.submission.data.textArea).to.equal(
+            form.submission.data.textField + form.submission.metadata.timezone
+          );
+
+          done();
         }, 250);
       }).catch(done);
     });
@@ -2780,49 +3169,46 @@ describe('Webform tests', function() {
     }).catch(done);
   }).timeout(3000);
 
-  //TOFIX
-  // it('Should have number and currency fields in empty form submission', function(done) {
-  //   const formElement = document.createElement('div');
-  //   const form= new Webform(formElement);
-  //   const formJson = {
-  //     components: [
-  //       {
-  //         label: 'Number',
-  //         key: 'number',
-  //         type: 'number'
-  //       },
-  //       {
-  //         label: 'Currency',
-  //         key: 'currency',
-  //         type: 'currency'
-  //       },
-  //       {
-  //         type: 'button',
-  //         label: 'Submit',
-  //         key: 'submit'
-  //       },
-  //     ],
-  //   };
+  it('Should have number and currency fields in empty form submission', function(done) {
+    const formElement = document.createElement('div');
+    const form= new Webform(formElement);
+    const formJson = {
+      components: [
+        {
+          label: 'Number',
+          key: 'number',
+          type: 'number'
+        },
+        {
+          label: 'Currency',
+          key: 'currency',
+          type: 'currency'
+        },
+        {
+          type: 'button',
+          label: 'Submit',
+          key: 'submit'
+        },
+      ],
+    };
 
-  //   const emptySubmissionData = {
-  //     number: '',
-  //     currency: '',
-  //     submit: true
-  //   };
+    const emptySubmissionData = {
+      submit: true
+    };
 
-  //   form.setForm(formJson).then(() => {
-  //     const clickEvent = new Event('click');
-  //     const submitBtn = form.element.querySelector('[name="data[submit]"]');
+    form.setForm(formJson).then(() => {
+      const clickEvent = new Event('click');
+      const submitBtn = form.element.querySelector('[name="data[submit]"]');
 
-  //     submitBtn.dispatchEvent(clickEvent);
+      submitBtn.dispatchEvent(clickEvent);
 
-  //     setTimeout(() => {
-  //       assert.deepEqual(form.data, emptySubmissionData);
-  //       done();
-  //     }, 200);
-  //   })
-  //   .catch((err) => done(err));
-  // });
+      setTimeout(() => {
+        assert.deepEqual(form.data, emptySubmissionData);
+        done();
+      }, 400);
+    })
+    .catch((err) => done(err));
+  });
 
   it('Test Truncate Multiple Spaces', (done) => {
     const formElement = document.createElement('div');