+++ src/components/_classes/component/editForm/Component.edit.conditional.js
@@ -7,6 +7,7 @@ export default [
     title: 'Simple',
     key: 'simple-conditional',
     theme: 'default',
+    weight: 105,
     components: [
       {
         type: 'select',