+++ src/components/_classes/list/ListComponent.js
@@ -50,7 +50,7 @@ export default class ListComponent extends Field {
     };
     const template = this.sanitize(
       this.component.template
-        ? this.interpolate(this.component.template, { item: data })
+        ? this.interpolate(this.component.template, { item: data }, options)
         : data.label,
       this.shouldSanitizeValue,
     );