+++ src/components/datetime/DateTime.js
@@ -89,7 +89,6 @@ export default class DateTimeComponent extends Input {
       type: 'calendar',
       timezone,
       displayInTimezone: _.get(this.component, 'displayInTimezone', 'viewer'),
-      submissionTimezone: this.submissionTimezone,
       locale: this.options.language,
       useLocaleSettings: _.get(this.component, 'useLocaleSettings', false),
       allowInput: _.get(this.component, 'allowInput', true),
@@ -113,13 +112,6 @@ export default class DateTimeComponent extends Input {
 
     // Add the validators date.
     this.validators.push('date');
-  }
-
-  performInputMapping(input) {
-    if (input.widget && input.widget.settings) {
-      input.widget.settings.submissionTimezone = this.submissionTimezone;
-    }
-    return input;
   }
 
   get defaultSchema() {
@@ -176,10 +168,12 @@ export default class DateTimeComponent extends Input {
   }
 
   getValueAsString(value) {
-    const format = FormioUtils.convertFormatToMoment(this.component.format);
-    if (value && !this.attached && this.submissionTimezone) {
-      return FormioUtils.momentDate(value, format, this.submissionTimezone).format(format);
+    let format = FormioUtils.convertFormatToMoment(this.component.format);
+    format += format.match(/z$/) ? '' : ' z';
+    const timezone = this.timezone;
+    if (value && !this.attached && timezone) {
+      return _.trim(FormioUtils.momentDate(value, format, timezone).format(format));
     }
-    return (value ? moment(value).format(format) : value) || '';
+    return (value ? _.trim(moment(value).format(format)) : value) || '';
   }
 }