+++ src/components/datetime/DateTime.unit.js
@@ -105,9 +105,7 @@ describe('DateTime Component', () => {
       input.dispatchEvent(blurEvent);
 
       setTimeout(() => {
-        // assert.equal(dateTime.getValue().startsWith(expectedValueStart), true);
         assert.equal(input.value, expectedValue);
-console.log(22222, input.value, dateTime.getValue());
         document.innerHTML = '';
         done();
       }, 300);
@@ -678,7 +676,6 @@ console.log(22222, input.value, dateTime.getValue());
     }).catch(done);
   });
 
-  //TOFIX
   // it('Should provide correct date in selected timezone after submission', (done) => {
   //   const form = _.cloneDeep(comp9);
   //   const element = document.createElement('div');